move genesis test to own file
